applications/frontend/requirements.txt
python-fasthtml>=0.8.0
uvicorn>=0.30
httpx==0.25.2
